LLaVA COCO Finetune/requirements.txt
transformers>=4.35.3
datasets==2.16.0
peft
bitsandbytes
accelerate
pillow
hf_transfer
pycocotools